Run setup changed to new selection method and new UI.

+++ src/3D/OpenGLImage.cpp
@@ -199,8 +199,8 @@ void OpenGLImage::SetAlphaColor(unsigned char red, unsigned char green,
 		unsigned char blue)
 {
 	if(!HasAlpha()) InitAlpha();
-	for(size_t i = 0; i < GetWidth(); i++){
-		for(size_t j = 0; j < GetHeight(); j++){
+	for(int i = 0; i < GetWidth(); i++){
+		for(int j = 0; j < GetHeight(); j++){
 			if(GetRed(i, j) == red && GetGreen(i, j) == green
 					&& GetBlue(i, j) == blue){
 				SetAlpha(i, j, 0);
